+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Functionality to generate data for the related items component
+
 ## 1.0.0
 
 * Functionality to generate data for the breadcrumb component